Add a 'mkv' profile
This will generate a Matroska file containing a VP8 video track and an
AC3 audio track (or whatever the format the audio is in the VOB).

It uses ffmpeg to encode the video into VP8 and then mkvmerge to create
the matroska file while extracting the first audio track from the VOB
and muxing it in.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>