Add the ability to specify the VOB audio track to use
When using the MKV profile by default we take the first audio track
from the VOB file to mux in. Add a -a option so you can specify the
audio track id (as shown by mkvmerge -i for example) to use instead. E.g
to specify the second audio track

    $ encode_vobs -P mkv -a 2 -t 2 *.vob

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>