Allow to provide a custom encode command
This adds a new 'custom' profile that allows a custom encode command to
be passed via the -c argument switch. This should be the full command
that you want to have used on the input file(s).

It has two placeholders that must be specified, %i for the input file
(this will be replaced as is) and %o for the output file. %o will be the
filename without the extension so you should provide that also, e.g

    -P custom -c "ffmpeg -i %i %o.webm"

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>